refact(store): clean up old store
