+++ pkg/store/store.go
@@ -17,6 +17,10 @@ limitations under the License.
 package store
 
 import (
+	"time"
+
+	"github.com/caicloud/cyclone/pkg/log"
+	"github.com/caicloud/cyclone/pkg/wait"
 	"gopkg.in/mgo.v2"
 )
 
@@ -27,11 +31,16 @@ const (
 	pipelineCollectionName       string = "pipelines"
 	pipelineRecordCollectionName string = "pipelineRecords"
 	eventCollectionName          string = "events"
+
+	socketTimeout  = time.Second * 5
+	syncTimeout    = time.Second * 5
+	tickerDuration = time.Second * 5
 )
 
 var (
 	session *mgo.Session
 	saltKey string
+	mclosed chan struct{}
 )
 
 // DataStore is the type for mongo db store.
@@ -48,9 +57,32 @@ type DataStore struct {
 }
 
 // Init store mongo client session
-func Init(s *mgo.Session, key string) {
-	session = s
+// Init store mongo client session
+func Init(host string, gracePeriod time.Duration, closing chan struct{}, key string) (chan struct{}, error) {
 	saltKey = key
+	mclosed = make(chan struct{})
+	var err error
+
+	// dail mongo session
+	// wait mongodb set up
+	wait.Poll(time.Second, gracePeriod, func() (bool, error) {
+		session, err = mgo.Dial(host)
+		return err == nil, nil
+	})
+
+	if err != nil {
+		log.Errorf("Unable connect to mongodb addr %s", host)
+		return nil, err
+	}
+
+	log.Infof("connect to mongodb addr: %s", host)
+	// Set the session mode as Eventual to ensure that the socket is created for each request.
+	// Can switch to other mode only after the old APIs are cleaned up.
+	session.SetMode(mgo.Eventual, true)
+
+	go backgroundMongo(closing)
+
+	return mclosed, nil
 }
 
 // NewStore copy a mongo client session
@@ -75,4 +107,25 @@ func (d *DataStore) Close() {
 // Ping ping mongo server
 func (d *DataStore) Ping() error {
 	return d.s.Ping()
+}
+
+// Background goroutine for mongo. It can hold mongo connection & close session when progress exit.
+func backgroundMongo(closing chan struct{}) {
+	ticker := time.NewTicker(tickerDuration)
+	for {
+		select {
+		case <-ticker.C:
+			if err := session.Ping(); err != nil {
+				log.Errorf("Ping Mongodb with error %s", err.Error())
+				session.Refresh()
+				session.SetSocketTimeout(socketTimeout)
+				session.SetSyncTimeout(syncTimeout)
+			}
+		case <-closing:
+			session.Close()
+			log.Info("Mongodb session has been closed")
+			close(mclosed)
+			return
+		}
+	}
 }
